+++ src/plugin/init/utils/__tests__/fixtures/manifest.json
