+++ src/plugin/init/utils/__tests__/helpers/baseManifest.ts
