+++ src/plugin/init/utils/__tests__/manifest.test.ts
