+++ src/plugin/init/utils/__tests__/privateKey.test.ts
