+++ src/plugin/init/utils/__tests__/template.test.ts
