+++ src/plugin/init/utils/deps.ts
