+++ src/plugin/init/utils/lang.ts
