+++ src/plugin/init/utils/manifest.ts
