+++ src/plugin/init/utils/privateKey.ts
